mv /etc/sonic.cfg /run/sonic/sonic.cfg
When using this container in docker-compose it is not that easy to bind mount a file into the container. It has a way better user experience when one can just specify a folder. Currently this doesn't work with the default configuration as the sonic.cfg is directly within /etc.
By replacing /etc/sonic.cfg with a symlink to /run/sonic/sonic.cfg also avoid introducing a breaking change.